ProductsController was broken down to partial classes, and some namespace erros were corrected
