+++ Controllers/ProductsController.Shared.cs
@@ -1,0 +1,13 @@
+using DurgerKing.Data;
+
+namespace DurgerKing.Controllers;
+
+public partial class ProductsController
+{
+    private readonly IAppDbContext dbContext;
+
+    public ProductsController(IAppDbContext dbContext)
+    {
+        this.dbContext = dbContext;
+    }
+}
